requirements.txt
requests
beautifulsoup4
pandas
faiss-cpu
sentence-transformers
transformers
torch
accelerate
gradio
flask
flask-cors
